// app/dashboard/interfaces/Stats.ts
export interface Stats {
    isAuthenticated: boolean;
    user: {
      id: string;
      username: string;
      email: string;
    };
    publicationStats: {
      userPub1Percent: number;
      userPub2Percent: number;
      crowdPub1Percent: number;
      crowdPub2Percent: number;
      totalRatingsCount: number;
      pub1RatingsCount: number;
      pub2RatingsCount: number;
    };
    pub_1_crowd_total_bias: number;
    pub_2_crowd_total_bias: number;
    pub_1_personal_bias: {
      publication: string;
      total_bias: number;
      sensationalism: number;
      undue_weight_bias: number;
      speculative_content: number;
      tonality_bias: number;
      concision_bias: number;
      coverage_bias: number;
      distortion_bias: number;
      partisan_bias: number;
      favors_or_attacks: number;
      content_bias: number;
      structural_bias: number;
      gatekeeping_bias: number;
      decision_making_bias: number;
      mainstream_bias: number;
      false_balance_bias: number;
    };
    pub_2_personal_bias: {
      publication: string;
      total_bias: number;
      sensationalism: number;
      undue_weight_bias: number;
      speculative_content: number;
      tonality_bias: number;
      concision_bias: number;
      coverage_bias: number;
      distortion_bias: number;
      partisan_bias: number;
      favors_or_attacks: number;
      content_bias: number;
      structural_bias: number;
      gatekeeping_bias: number;
      decision_making_bias: number;
      mainstream_bias: number;
      false_balance_bias: number;
    };
    pub_1_crowd_bias: {
      publication: string;
      total_bias: number;
      sensationalism: number;
      undue_weight_bias: number;
      speculative_content: number;
      tonality_bias: number;
      concision_bias: number;
      coverage_bias: number;
      distortion_bias: number;
      partisan_bias: number;
      favors_or_attacks: number;
      content_bias: number;
      structural_bias: number;
      gatekeeping_bias: number;
      decision_making_bias: number;
      mainstream_bias: number;
      false_balance_bias: number;
    };
    pub_2_crowd_bias: {
      publication: string;
      total_bias: number;
      sensationalism: number;
      undue_weight_bias: number;
      speculative_content: number;
      tonality_bias: number;
      concision_bias: number;
      coverage_bias: number;
      distortion_bias: number;
      partisan_bias: number;
      favors_or_attacks: number;
      content_bias: number;
      structural_bias: number;
      gatekeeping_bias: number;
      decision_making_bias: number;
      mainstream_bias: number;
      false_balance_bias: number;
    };
  }